+++ esp32h2-hal/examples/spi_eh1_loopback.rs
@@ -22,7 +22,10 @@ use esp32h2_hal::{
     gpio::IO,
     peripherals::Peripherals,
     prelude::*,
-    spi::{master::Spi, SpiMode},
+    spi::{
+        master::{prelude::*, Spi},
+        SpiMode,
+    },
     Delay,
 };
 use esp_backtrace as _;