+++ esp32s3-hal/examples/spi_eh1_device_loopback.rs
@@ -25,7 +25,7 @@ use esp32s3_hal::{
     peripherals::Peripherals,
     prelude::*,
     spi::{
-        master::{Spi, SpiBusController},
+        master::{prelude::*, Spi, SpiBusController},
         SpiMode,
     },
     Delay,